+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/config/VirtualCluster.java
@@ -5,6 +5,7 @@
  */
 package io.kroxylicious.proxy.config;
 
+import java.util.List;
 import java.util.Optional;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -13,21 +14,30 @@ import io.kroxylicious.proxy.config.tls.Tls;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProvider;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProviderService;
 
+import edu.umd.cs.findbugs.annotations.NonNull;
+import edu.umd.cs.findbugs.annotations.Nullable;
+
 public record VirtualCluster(TargetCluster targetCluster,
                              @JsonProperty(required = true) ClusterNetworkAddressConfigProviderDefinition clusterNetworkAddressConfigProvider,
-
                              @JsonProperty() Optional<Tls> tls,
                              boolean logNetwork,
-                             boolean logFrames) {
-    public io.kroxylicious.proxy.model.VirtualCluster toVirtualClusterModel(PluginFactoryRegistry pfr, String virtualClusterNodeName) {
+                             boolean logFrames,
+                             @Nullable List<String> filters) {
+
+    public io.kroxylicious.proxy.model.VirtualCluster toVirtualClusterModel(@NonNull PluginFactoryRegistry pfr,
+                                                                            @NonNull List<NamedFilterDefinition> filterDefinitions,
+                                                                            @NonNull String virtualClusterNodeName) {
+
         return new io.kroxylicious.proxy.model.VirtualCluster(virtualClusterNodeName,
                 targetCluster(),
                 toClusterNetworkAddressConfigProviderModel(pfr),
                 tls(),
-                logNetwork(), logFrames());
+                logNetwork(),
+                logFrames(),
+                filterDefinitions);
     }
 
-    private ClusterNetworkAddressConfigProvider toClusterNetworkAddressConfigProviderModel(PluginFactoryRegistry registry) {
+    private ClusterNetworkAddressConfigProvider toClusterNetworkAddressConfigProviderModel(@NonNull PluginFactoryRegistry registry) {
         ClusterNetworkAddressConfigProviderService provider = registry.pluginFactory(ClusterNetworkAddressConfigProviderService.class)
                 .pluginInstance(clusterNetworkAddressConfigProvider.type());
         return provider.build(clusterNetworkAddressConfigProvider.config());